+++ test/fixtures/person.rb
@@ -2,10 +2,10 @@
 
 require "structure"
 
-# integer, boolean, simple predicate
+# integer, boolean, simple predicate, optional attribute
 Person = Structure.new do
   attribute(:name, String)
-  attribute(:age, Integer)
+  attribute?(:age, Integer)
   attribute(:active, :boolean, default: true)
   attribute(:email, String, from: "email_address")
 end